pyproject.toml
[project]
name = "oauth2-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
]
